Fix entry info when cached
Before:
When caching is turned on, the user_info was empty in some cases, including inside a view. The problem was that the FrmEntry::getAll function added to the cache for FrmEntry::getOne, but the query wasn't exactly the same.

After:
Now if an entry is cached from the FrmEntry::getAll function, it will include "description" like the FrmEntry::getOne function.